Add order tab test and refactor use of test data helper
