+++ Tests/Unit/Controller/Admin/Transaction/TransactionTabTest.php
@@ -8,7 +8,7 @@
  */
 
 use Wirecard\Oxid\Controller\Admin\Transaction\TransactionTab;
-use Wirecard\Oxid\Tests\Unit\Controller\Admin\Transaction\TransactionTestHelper;
+use Wirecard\Oxid\Tests\Unit\Controller\Admin\TestDataHelper;
 
 class TransactionTabTest extends Wirecard\Test\WdUnitTestCase
 {
@@ -19,7 +19,7 @@ class TransactionTabTest extends Wirecard\Test\WdUnitTestCase
 
     protected function dbData()
     {
-        return TransactionTestHelper::getDemoData();
+        return TestDataHelper::getDemoData();
     }
 
     public function testGetListData()
